Improve menu_cmd CmdDraw matching (float h for single-precision ops)
Declare the box height 'h' as float instead of double so the t+=h and
h+=smallOffset arithmetic emits fadds/fsubs matching the target's
single-precision idiom (was forcing fadd + double-pool loads).
CmdDraw 68.64->68.88%, unit 75.80->75.86%.

+++ src/menu_cmd.cpp
@@ -978,7 +978,7 @@ void CMenuPcs::CmdDraw()
 			const float x = static_cast<float>(entry->x);
 			float y = static_cast<float>(entry->y);
 			const float w = static_cast<float>(entry->width);
-			double h = static_cast<double>(entry->height);
+			float h = static_cast<float>(entry->height);
 			const float u = entry->u;
 			float t = smallOffset;
